# server/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "stream.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/admin": {
      "get": {
        "description": "Admin index page",
        "tags": [
          "admin"
        ],
        "operationId": "adminIndex",
        "responses": {
          "200": {
            "$ref": "#/responses/htmlPage"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
    "/admin/add_user": {
      "post": {
        "description": "Add user",
        "tags": [
          "admin",
          "user"
        ],
        "operationId": "addUser",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Login",
            "name": "login",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Password",
            "name": "password",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "boolean",
            "x-go-name": "IsAdmin",
            "name": "is_admin",
            "in": "query"
          }
        ],
        "responses": {
          "303": {
            "$ref": "#/responses/seeOther"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
    "/admin/change_user": {
      "post": {
        "tags": [
          "admin",
          "user"
        ],
        "summary": "Change user's data.",
        "operationId": "changeUser",
        "responses": {
          "200": {
            "$ref": "#/responses/statusOk"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
    "/admin/delete_user": {
      "delete": {
        "tags": [
          "admin",
          "user"
        ],
        "summary": "Completely delete user.",
        "operationId": "deleteUser",
        "responses": {
          "303": {
            "$ref": "#/responses/seeOther"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
    "/admin/login": {
      "get": {
        "description": "Uses `LoginForm` struct to deserialize.",
        "tags": [
          "admin"
        ],
        "summary": "Gets from with login and hashed password.",
        "operationId": "adminLogin",
        "responses": {
          "200": {
            "$ref": "#/responses/statusOk"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
    "/admin/songs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Shows list of all songs.",
        "operationId": "adminSongs",
        "responses": {
          "200": {
            "$ref": "#/responses/htmlPage"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "description": "Uses `LoginForm` struct to deserialize.",
        "tags": [
          "user"
        ],
        "summary": "Gets json with login and hashed password.",
        "operationId": "userLogin",
        "parameters": [
          {
            "x-go-name": "Login",
            "description": "Login",
            "name": "login",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Password",
            "description": "Password",
            "name": "password",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/statusOk"
          },
          "400": {
            "$ref": "#/responses/badRequest"
          },
          "500": {
            "$ref": "#/responses/internalServerError"
          }
        }
      }
    },
...
  },
  "responses": {
    "badRequest": {
      "description": "",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Message"
        },
        "message": {
          "type": "string",
          "description": "Code"
        }
      }
    },
...
    "htmlPage": {
      "description": "",
      "headers": {
        "data": {
          "type": "string"
        }
      }
    },
    "internalServerError": {
      "description": "",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Message"
        },
        "message": {
          "type": "string",
          "description": "Code"
        }
      }
    },
...
    "seeOther": {
      "description": "",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Message"
        },
        "message": {
          "type": "string",
          "description": "Code"
        }
      }
    },
...
    "statusOk": {
      "description": "",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Message"
        },
        "message": {
          "type": "string",
          "description": "Code"
        }
      }
    }
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}